Move internal id tests into separate files
