+++ webdriver/tests/bidi/script/evaluate/internal_id.py
@@ -1,0 +1,40 @@
+import pytest
+
+from ... import recursive_compare, any_string
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize(
+    "return_structure, result_type",
+    [
+        ("[data, data]", "array"),
+        ("new Map([['foo', data],['bar', data]])", "map"),
+        ("({ 'foo': data, 'bar': data })", "object"),
+    ],
+)
+@pytest.mark.parametrize(
+    "expression, type",
+    [
+        ("[1]", "array"),
+        ("new Map([[true, false]])", "map"),
+        ("new Set(['baz'])", "set"),
+        ("{ baz: 'qux' }", "object"),
+    ],
+)
+async def test_remote_values_with_internal_id(
+    evaluate, return_structure, result_type, expression, type
+):
+    result = await evaluate(f"{{const data = {expression}; {return_structure}}}")
+
+    if result_type == "array":
+        value = [
+            {"type": type, "internalId": any_string},
+            {"type": type, "internalId": any_string},
+        ]
+    else:
+        value = [
+            ["foo", {"type": type, "internalId": any_string}],
+            ["bar", {"type": type, "internalId": any_string}],
+        ]
+
+    recursive_compare({"type": result_type, "value": value}, result)
